Python Challenge 1 and 2
Given a target number and a txt file containing a large number of numbers, find the indices of 2 numbers in the original list that add up to the target. 

V1 uses sorted() to sort the entire list an improvised attempt at a binary sort. 

V2 uses an insertion sort to only sort numbers in the list that are less than the target, followed by a properly-structured recursive binary search to locate the addends.